+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # 0.2.6
 
++ Remove `GenericCaseTree`, add `GenericCase` because now each case in the
+  case tree have their own telescope
+
 # 0.2.5
 
 + Rename `U` to `Type`, `sum` to `Sum`